FullyConnectedTensorProduct to haiku submodule
