+++ e3nn_jax/__init__.py
@@ -44,7 +44,6 @@ from e3nn_jax._src.linear_haiku import Linear  # TODO remove
 from e3nn_jax._src.core_tensor_product import FunctionalTensorProduct
 from e3nn_jax._src.tensor_products import (
     FunctionalFullyConnectedTensorProduct,
-    FullyConnectedTensorProduct,
     tensor_product,
     elementwise_tensor_product,
     tensor_square,
@@ -59,6 +58,7 @@ from e3nn_jax._src.graph_util import index_add, radius_graph, scatter_sum
 from e3nn_jax._src.reduced_tensor_product import reduced_tensor_product_basis, reduced_symmetric_tensor_product_basis
 from e3nn_jax._src.symmetric_tensor_product import SymmetricTensorProduct  # TODO remove
 from e3nn_jax._src.s2grid import from_s2grid, to_s2grid, s2grid
+from e3nn_jax._src.fc_tp_haiku import FullyConnectedTensorProduct  # TODO remove
 
 # make submodules flax and haiku available
 from e3nn_jax import flax, haiku
